changing class names

+++ src/root/tests/jive/pattern.cpp
@@ -53,7 +53,7 @@ Y_UTEST(jive_pattern)
     Y_SIZEOF(Jive::Any1);
 
     { Jive::Any1      p;          testPattern(p); }
-    { Jive::Single    p('A');     testPattern(p); }
+    { Jive::Byte      p('A');     testPattern(p); }
     { Jive::Range     p('Z','A'); testPattern(p); }
     { Jive::Excluded  p('X');     testPattern(p); }
